reorganizes fixtures conftests
